Moved functions to TimeUtil.kt
