+++ coding-projects/android/TaskTracker/app/src/main/java/com/example/tasktracker/TimeUtil.kt
@@ -20,5 +20,16 @@ class TimeUtil {
             val day = SimpleDateFormat("EE", Locale.getDefault()).format(c)
             return day == dayOfWeek
         }
+
+        fun convertMillisToDate(millis: Long): String {
+            val formatter = SimpleDateFormat("dd MMM yyyy", Locale.getDefault())
+            formatter.timeZone = TimeZone.getTimeZone("UTC")
+            return formatter.format(Date(millis))
+        }
+
+        fun convertTime(time: Date): String {
+            val formatter = SimpleDateFormat("HH:mm")
+            return formatter.format(time).toString()
+        }
     }
 }
